Make AddPeer.fromProto generic
